refactor: split daemon.ts into focused modules
Extract five focused modules from the 1,883-line daemon.ts:

- hook-summary.ts     — extractHookSummary() for web UI hook display
- runner-state.ts     — process-lock + persistent identity + isPidRunning()
- runner-usage-cache.ts — provider quota cache (Anthropic/Gemini/Codex)
- workspace.ts        — getWorkspaceRoots() + isCwdAllowed() path guards
- session-spawner.ts  — RunnerSession + spawnSession() worker management

daemon.ts shrinks from 1,883 → ~980 lines, retaining only runDaemon()
and its Socket.IO event handlers.

Callers updated:
- stop.ts: daemon.js → runner-state.js
- supervisor.ts: daemon.js → runner-state.js
- daemon-hooks.test.ts: daemon.js → hook-summary.js

No logic changes — pure reorganization. Tests pass, typecheck clean.
