+++ packages/cli/src/runner/daemon.ts
@@ -1,12 +1,10 @@
-import { spawn, exec, execFile, execFileSync, type ChildProcess } from "node:child_process";
+import { exec, execFile } from "node:child_process";
 import { promisify } from "node:util";
 
 const execAsync = promisify(exec);
 const execFileAsync = promisify(execFile);
-import { existsSync, mkdirSync, readdirSync, readFileSync, realpathSync, rmSync, statSync, writeFileSync } from "node:fs";
-import { mkdir, readdir, stat, readFile } from "node:fs/promises";
-import { randomBytes, randomUUID } from "node:crypto";
-import { homedir, hostname } from "node:os";
+import { readdir, stat } from "node:fs/promises";
+import { hostname } from "node:os";
 import {
     spawnTerminal,
     writeTerminalInput,
@@ -15,196 +13,17 @@ import {
     listTerminals,
     killAllTerminals,
 } from "./terminal.js";
-import { join, dirname, relative, basename, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
+import { join, basename } from "node:path";
 import { io, type Socket } from "socket.io-client";
-import type { RunnerClientToServerEvents, RunnerServerToClientEvents, RunnerHook } from "@pizzapi/protocol";
-import { loadGlobalConfig, loadConfig, defaultAgentDir, expandHome, type HooksConfig } from "../config.js";
-import { AuthStorage } from "@mariozechner/pi-coding-agent";
+import type { RunnerClientToServerEvents, RunnerServerToClientEvents } from "@pizzapi/protocol";
+import { loadGlobalConfig } from "../config.js";
 import { cleanupSessionAttachments, sweepOrphanedAttachments } from "../extensions/session-attachments.js";
-import { getRefreshedOAuthToken, parseGeminiQuotaCredential } from "./usage-auth.js";
-import { setLogComponent, logInfo, logWarn, logError, logAuth } from "./logger.js";
-
-/**
- * Summarise the active hooks from a HooksConfig for display in the web UI.
- * Returns one entry per hook type that has at least one configured command.
- * Scripts are represented by the basename of the first token in the command string.
- */
-export function extractHookSummary(hooks?: HooksConfig): RunnerHook[] {
-    if (!hooks) return [];
-    const result: RunnerHook[] = [];
-
-    // Matcher-based hooks (PreToolUse / PostToolUse)
-    for (const type of ["PreToolUse", "PostToolUse"] as const) {
-        const matchers = hooks[type] ?? [];
-        const scripts: string[] = [];
-        for (const m of matchers) {
-            for (const h of m.hooks) {
-                const cmd = h.command.trim().split(/\s+/)[0];
-                if (cmd) scripts.push(basename(cmd));
-            }
-        }
-        if (scripts.length > 0) result.push({ type, scripts });
-    }
-
-    // Entry-based hooks
-    const entryTypes = [
-        "Input",
-        "BeforeAgentStart",
-        "UserBash",
-        "SessionBeforeSwitch",
-        "SessionBeforeFork",
-        "SessionShutdown",
-        "SessionBeforeCompact",
-        "SessionBeforeTree",
-        "ModelSelect",
-    ] as const;
-    for (const type of entryTypes) {
-        const entries = (hooks[type] as { command: string }[] | undefined) ?? [];
-        const scripts: string[] = [];
-        for (const h of entries) {
-            const cmd = h.command.trim().split(/\s+/)[0];
-            if (cmd) scripts.push(basename(cmd));
-        }
-        if (scripts.length > 0) result.push({ type, scripts });
-    }
-
-    return result;
-}
-
-interface RunnerSession {
-    sessionId: string;
-    child: ChildProcess | null;
-    startedAt: number;
-    /**
-     * True if this session was re-adopted after a daemon restart.
-     * Adopted sessions have no child process handle — the worker is still
-     * running independently with its own relay connection.
-     */
-    adopted?: boolean;
-    /** ID of the parent session that spawned this one. */
-    parentSessionId?: string;
-}
-
-// ── Runner state file (~/.pizzapi/runner.json) ────────────────────────────────
-//
-// A single JSON file consolidates both the process-lock (pid + startedAt) and
-// the persistent runner identity (runnerId + runnerSecret).  This keeps all
-// runner state in the canonical ~/.pizzapi/ directory alongside config.json.
-//
-// Schema:
-//   {
-//     "pid": 12345,            // PID of the currently-running daemon (lock)
-//     "supervisorPid": 12344,  // PID of the outer supervisor process
-//     "startedAt": "<iso>",    // ISO timestamp of that daemon start
-//     "runnerId": "<uuid>",    // stable runner identity (never changes)
-//     "runnerSecret": "<hex>"  // 32-byte secret used to re-authenticate
-//   }
-
-interface RunnerState {
-    pid: number;
-    supervisorPid?: number;
-    startedAt: string;
-    runnerId: string;
-    runnerSecret: string;
-}
-
-export function defaultStatePath(): string {
-    return process.env.PIZZAPI_RUNNER_STATE_PATH ?? join(homedir(), ".pizzapi", "runner.json");
-}
-
-/**
- * Acquire the runner lock and load (or create) the persistent identity.
- * Both live in a single JSON file so they stay in sync atomically.
- *
- * Returns the identity portion on success; exits the process if another
- * live runner already holds the lock.
- */
-function acquireStateAndIdentity(statePath: string): { runnerId: string; runnerSecret: string } {
-    // Ensure the parent directory exists.
-    const dir = join(statePath, "..");
-    try {
-        mkdirSync(dir, { recursive: true });
-    } catch {
-        // ignore — already exists or unwritable (caught below)
-    }
-
-    // Attempt up to two passes: one normal, one after clearing a stale lock.
-    for (let attempt = 0; attempt < 2; attempt++) {
-        let existing: Partial<RunnerState> = {};
-        if (existsSync(statePath)) {
-            try {
-                existing = JSON.parse(readFileSync(statePath, "utf-8")) as Partial<RunnerState>;
-            } catch {
-                // Corrupt file — treat as empty, overwrite below.
-            }
-
-            // Check whether another live daemon holds the lock.
-            const pid = typeof existing.pid === "number" ? existing.pid : NaN;
-            if (Number.isFinite(pid) && pid > 0) {
-                if (isPidRunning(pid)) {
-                    logError(`pizzapi runner already running (pid ${pid}, state: ${statePath}).`);
-                    logError("   Stop the existing runner process first, e.g.: kill ${pid}");
-                    process.exit(1);
-                }
-                // PID is gone or belongs to an unrelated process — stale lock.
-                logInfo(`clearing stale lock (pid ${pid} is no longer a runner process)`);
-            }
-        }
-
-        // Write the new lock (preserving identity if already present).
-        const runnerId =
-            typeof existing.runnerId === "string" && existing.runnerId.length > 0
-                ? existing.runnerId
-                : randomUUID();
-        const runnerSecret =
-            typeof existing.runnerSecret === "string" && existing.runnerSecret.length > 0
-                ? existing.runnerSecret
-                : randomBytes(32).toString("hex");
-
-        const state: RunnerState = {
-            pid: process.pid,
-            supervisorPid: typeof existing.supervisorPid === "number" ? existing.supervisorPid : undefined,
-            startedAt: new Date().toISOString(),
-            runnerId,
-            runnerSecret,
-        };
-
-        try {
-            writeFileSync(statePath, JSON.stringify(state, null, 2), { encoding: "utf-8", mode: 0o600 });
-            return { runnerId, runnerSecret };
-        } catch (err: any) {
-            logError(`Failed to write runner state to ${statePath}: ${err?.message ?? String(err)}`);
-            process.exit(1);
-        }
-    }
-
-    // Should never reach here.
-    process.exit(1);
-}
-
-/**
- * Release the process lock by clearing the pid field in the state file,
- * while preserving the persistent identity for the next run.
- */
-function releaseStateLock(statePath: string) {
-    try {
-        const existing = JSON.parse(readFileSync(statePath, "utf-8")) as Partial<RunnerState>;
-        // Only clear the lock fields; keep runnerId + runnerSecret intact.
-        const updated = {
-            pid: 0,
-            supervisorPid: 0,
-            startedAt: "",
-            runnerId: existing.runnerId ?? "",
-            runnerSecret: existing.runnerSecret ?? "",
-        };
-        writeFileSync(statePath, JSON.stringify(updated, null, 2), { encoding: "utf-8", mode: 0o600 });
-    } catch {
-        // Best-effort — ignore errors on shutdown.
-    }
-}
-
-// ── Skill helpers (re-exported from shared module) ────────────────────────────
+import { setLogComponent, logInfo, logWarn, logError } from "./logger.js";
+import { extractHookSummary } from "./hook-summary.js";
+import { defaultStatePath, acquireStateAndIdentity, releaseStateLock } from "./runner-state.js";
+import { startUsageRefreshLoop, stopUsageRefreshLoop } from "./runner-usage-cache.js";
+import { getWorkspaceRoots, isCwdAllowed } from "./workspace.js";
+import { type RunnerSession, spawnSession } from "./session-spawner.js";
 
 import {
     type SkillMeta,
@@ -224,7 +43,6 @@ import {
 
 import {
     scanAllPluginInfo,
-    type PluginInfo,
 } from "../plugins.js";
 
 /**
@@ -236,307 +54,6 @@ function resolveConfigRelayUrl(): string | undefined {
     const cfg = loadGlobalConfig();
     const url = cfg.relayUrl;
     return url && url !== "off" ? url : undefined;
-}
-
-// ── Runner-wide usage cache (shared with worker processes via file) ───────────
-//
-// The runner daemon is the single source of truth for provider quota data on
-// a given machine.  All worker sessions inherit PIZZAPI_RUNNER_USAGE_CACHE_PATH
-// and read from this file instead of each making their own API calls.
-
-interface UsageWindow { label: string; utilization: number; resets_at: string }
-interface ProviderUsageData {
-    windows: UsageWindow[];
-    status?: "ok" | "unknown";
-    errorCode?: number;
-}
-interface RunnerUsageCacheFile {
-    fetchedAt: number;
-    providers: Record<string, ProviderUsageData>;
-}
-
-/** Runner cache write cadence for provider usage snapshots. */
-const RUNNER_USAGE_REFRESH_INTERVAL = 5 * 60 * 1000;
-/** Anthropic usage changes slowly; poll less frequently by default. */
-const ANTHROPIC_USAGE_REFRESH_INTERVAL = 15 * 60 * 1000;
-
-let _usageRefreshTimer: ReturnType<typeof setInterval> | null = null;
-let _lastAnthropicUsage: { data: ProviderUsageData | null; fetchedAt: number } | null = null;
-
-/**
- * Tracks CWDs of active worker sessions so usage fetches can probe
- * project-local agentDir overrides when the daemon runs from a different directory.
- * Map: cwd → set of sessionIds using that cwd.
- */
-const _activeSessionCwds = new Map<string, Set<string>>();
-
-function trackSessionCwd(sessionId: string, cwd: string): void {
-    if (!_activeSessionCwds.has(cwd)) _activeSessionCwds.set(cwd, new Set());
-    _activeSessionCwds.get(cwd)!.add(sessionId);
-}
-
-function untrackSessionCwd(sessionId: string, cwd: string): void {
-    const sessions = _activeSessionCwds.get(cwd);
-    if (!sessions) return;
-    sessions.delete(sessionId);
-    if (sessions.size === 0) _activeSessionCwds.delete(cwd);
-}
-
-function runnerUsageCacheFilePath(): string {
-    return join(homedir(), ".pizzapi", "usage-cache.json");
-}
-
-/**
- * Returns all unique AuthStorage instances known to the daemon:
- * the daemon's own startup CWD first, followed by any CWD registered by active
- * worker sessions that maps to a different auth.json (e.g. a project-specific
- * agentDir override).  Results are deduplicated by resolved auth.json path so
- * the same file is never probed twice.
- *
- * Usage fetch functions iterate this list and use the first storage that
- * yields valid credentials, ensuring that sessions spawned in projects with
- * their own agentDir overrides are covered even when the daemon was started
- * from a different directory.
- */
-function getKnownAuthStorages(): AuthStorage[] {
-    const seen = new Set<string>();
-    const result: AuthStorage[] = [];
-    const cwds = [process.cwd(), ..._activeSessionCwds.keys()];
-    for (const cwd of cwds) {
-        const config = loadConfig(cwd);
-        const agentDir = config.agentDir ? expandHome(config.agentDir) : defaultAgentDir();
-        const authPath = join(agentDir, "auth.json");
-        if (!seen.has(authPath)) {
-            seen.add(authPath);
-            result.push(AuthStorage.create(authPath));
-        }
-    }
-    return result;
-}
-
-async function fetchAnthropicUsageData(): Promise<ProviderUsageData | null> {
-    let token: string | null = null;
-    try {
-        for (const authStorage of getKnownAuthStorages()) {
-            token = await getRefreshedOAuthToken(authStorage, "anthropic");
-            if (token) break;
-        }
-    } catch (err: any) {
-        logWarn(`failed to get Anthropic credentials: ${err?.message ?? String(err)}`);
-        return null;
-    }
-    if (!token) return null;
-    try {
-        const res = await fetch("https://api.anthropic.com/api/oauth/usage", {
-            headers: {
-                Authorization: `Bearer ${token}`,
-                "anthropic-version": "2023-06-01",
-                "anthropic-beta": "oauth-2025-04-20",
-            },
-        });
-        if (!res.ok) {
-            if (res.status === 403) return { windows: [], status: "unknown", errorCode: 403 };
-            return null;
-        }
-        const raw = (await res.json()) as Record<string, unknown>;
-        const WINDOW_LABELS: Record<string, string> = {
-            five_hour: "5-hour",
-            seven_day: "7-day",
-            seven_day_opus: "7-day (Opus)",
-            seven_day_sonnet: "7-day (Sonnet)",
-            seven_day_oauth_apps: "7-day (OAuth apps)",
-            seven_day_cowork: "7-day (co-work)",
-        };
-        const windows: UsageWindow[] = [];
-        for (const [key, label] of Object.entries(WINDOW_LABELS)) {
-            const w = raw[key] as { utilization: number; resets_at: string } | null | undefined;
-            if (w?.resets_at != null && typeof w.utilization === "number") {
-                windows.push({ label, utilization: w.utilization, resets_at: w.resets_at });
-            }
-        }
-        return windows.length > 0 ? { windows, status: "ok" } : null;
-    } catch {
-        return null;
-    }
-}
-
-async function getRunnerAnthropicUsageData(opts: { force?: boolean } = {}): Promise<ProviderUsageData | null> {
-    const now = Date.now();
-    const force = opts.force === true;
-
-    if (!force && _lastAnthropicUsage && now - _lastAnthropicUsage.fetchedAt < ANTHROPIC_USAGE_REFRESH_INTERVAL) {
-        return _lastAnthropicUsage.data;
-    }
-
-    const data = await fetchAnthropicUsageData();
-    // Only cache successful fetches so transient failures don't suppress
-    // retries for the full 15-minute interval.
-    if (data !== null) {
-        _lastAnthropicUsage = { data, fetchedAt: now };
-    }
-    return data;
-}
-
-async function fetchGeminiUsageData(): Promise<ProviderUsageData | null> {
-    let token: string | undefined;
-    let projectId: string | undefined;
-    try {
-        for (const authStorage of getKnownAuthStorages()) {
-            // AuthStorage.getApiKey handles OAuth token refresh and returns
-            // JSON.stringify({ token, projectId }) via the provider's getApiKey().
-            // Use parseGeminiQuotaCredential to validate the result — API-key
-            // credentials return a plain string that fails JSON.parse, so we
-            // must not short-circuit on the first truthy raw value; we need to
-            // confirm it is a valid OAuth Gemini credential before stopping.
-            const raw = await authStorage.getApiKey("google-gemini-cli");
-            const cred = parseGeminiQuotaCredential(raw);
-            if (cred) {
-                token = cred.token;
-                projectId = cred.projectId;
-                break;
-            }
-        }
-    } catch (err: any) {
-        logWarn(`failed to get Google credentials: ${err?.message ?? String(err)}`);
-        return null;
-    }
-    if (!token || !projectId) return null;
-    try {
-        const endpoint = process.env["CODE_ASSIST_ENDPOINT"] ?? "https://cloudcode-pa.googleapis.com";
-        const version = process.env["CODE_ASSIST_API_VERSION"] ?? "v1internal";
-        const res = await fetch(`${endpoint}/${version}:retrieveUserQuota`, {
-            method: "POST",
-            headers: { Authorization: `Bearer ${token}`, "Content-Type": "application/json" },
-            body: JSON.stringify({ project: projectId }),
-        });
-        if (!res.ok) {
-            if (res.status === 403) return { windows: [], status: "unknown", errorCode: 403 };
-            return null;
-        }
-        const raw = (await res.json()) as {
-            buckets?: Array<{ remainingFraction?: number; resetTime?: string; tokenType?: string; modelId?: string }>;
-        };
-        const windows: UsageWindow[] = [];
-        for (const bucket of raw.buckets ?? []) {
-            if (bucket.remainingFraction == null || !bucket.resetTime) continue;
-            const utilization = (1 - bucket.remainingFraction) * 100;
-            const label = [bucket.tokenType, bucket.modelId].filter(Boolean).join(" / ") || "Quota";
-            windows.push({ label, utilization, resets_at: bucket.resetTime });
-        }
-        return windows.length > 0 ? { windows, status: "ok" } : null;
-    } catch {
-        return null;
-    }
-}
-
-async function fetchCodexUsageData(): Promise<ProviderUsageData | null> {
-    let token: string | null = null;
-    try {
-        for (const authStorage of getKnownAuthStorages()) {
-            token = await getRefreshedOAuthToken(authStorage, "openai-codex");
-            if (token) break;
-        }
-    } catch (err: any) {
-        logWarn(`failed to get OpenAI Codex credentials: ${err?.message ?? String(err)}`);
-        return null;
-    }
-    if (!token) return null;
-    try {
-        const res = await fetch("https://chatgpt.com/backend-api/wham/usage", {
-            headers: { Authorization: `Bearer ${token}` },
-        });
-        if (!res.ok) {
-            if (res.status === 403) return { windows: [], status: "unknown", errorCode: 403 };
-            return null;
-        }
-        const raw = (await res.json()) as any;
-
-        function windowLabel(minutes: number | null | undefined): string {
-            if (!minutes) return "Usage";
-            if (minutes < 60) return `${minutes}-min`;
-            if (minutes < 60 * 24) return `${Math.round(minutes / 60)}-hour`;
-            return `${Math.round(minutes / 60 / 24)}-day`;
-        }
-        function toWin(w: any, label: string): UsageWindow | null {
-            if (!w) return null;
-            const used = typeof w.used_percent === "number" ? w.used_percent : null;
-            const resetAt = "resets_at" in w ? w.resets_at : "reset_at" in w ? w.reset_at : null;
-            if (used == null || resetAt == null) return null;
-            const minutes = "window_minutes" in w
-                ? (w.window_minutes ?? undefined)
-                : "limit_window_seconds" in w && typeof w.limit_window_seconds === "number"
-                    ? Math.max(1, Math.round(w.limit_window_seconds / 60))
-                    : undefined;
-            return {
-                label: minutes ? windowLabel(minutes) : label,
-                utilization: used,
-                resets_at: new Date(resetAt * 1000).toISOString(),
-            };
-        }
-
-        const windows: UsageWindow[] = [];
-        const primary = toWin(raw.rate_limit?.primary_window ?? raw.rate_limit?.primary, "Primary");
-        if (primary) windows.push(primary);
-        const secondary = toWin(raw.rate_limit?.secondary_window ?? raw.rate_limit?.secondary, "Secondary");
-        if (secondary) windows.push(secondary);
-        for (const extra of raw.additional_rate_limits ?? []) {
-            const w = toWin(extra.rate_limit?.primary_window ?? extra.rate_limit?.primary, extra.limit_name);
-            if (w) { w.label = extra.limit_name; windows.push(w); }
-        }
-        return windows.length > 0 ? { windows, status: "ok" } : null;
-    } catch {
-        return null;
-    }
-}
-
-/**
- * Fetch usage from all configured providers and write the result to the shared
- * cache file so every worker on this runner node can read it without making
- * their own API calls.
- */
-async function refreshAndWriteRunnerUsageCache(opts: { forceAnthropic?: boolean } = {}): Promise<void> {
-    const [anthropicResult, geminiResult, codexResult] = await Promise.allSettled([
-        getRunnerAnthropicUsageData({ force: opts.forceAnthropic === true }),
-        fetchGeminiUsageData(),
-        fetchCodexUsageData(),
-    ]);
-
-    const providers: Record<string, ProviderUsageData> = {};
-    if (anthropicResult.status === "fulfilled" && anthropicResult.value) {
-        providers.anthropic = anthropicResult.value;
-    }
-    if (geminiResult.status === "fulfilled" && geminiResult.value) {
-        providers["google-gemini-cli"] = geminiResult.value;
-    }
-    if (codexResult.status === "fulfilled" && codexResult.value) {
-        providers["openai-codex"] = codexResult.value;
-    }
-
-    if (Object.keys(providers).length === 0) return; // No credentials available — skip write
-
-    const cache: RunnerUsageCacheFile = { fetchedAt: Date.now(), providers };
-    try {
-        writeFileSync(runnerUsageCacheFilePath(), JSON.stringify(cache, null, 2), { encoding: "utf-8", mode: 0o600 });
-        logInfo(`usage cache refreshed (${Object.keys(providers).join(", ")})`);
-    } catch (err: any) {
-        logWarn(`failed to write usage cache: ${err?.message ?? String(err)}`);
-    }
-}
-
-function startUsageRefreshLoop(): void {
-    if (_usageRefreshTimer !== null) return;
-    // Kick off an immediate fetch so workers spawned right after startup have data.
-    void refreshAndWriteRunnerUsageCache();
-    _usageRefreshTimer = setInterval(() => {
-        void refreshAndWriteRunnerUsageCache();
-    }, RUNNER_USAGE_REFRESH_INTERVAL);
-}
-
-function stopUsageRefreshLoop(): void {
-    if (_usageRefreshTimer !== null) {
-        clearInterval(_usageRefreshTimer);
-        _usageRefreshTimer = null;
-    }
 }
 
 /**
@@ -897,6 +414,7 @@ export async function runDaemon(_args: string[] = []): Promise<number> {
                 sessions: Array.from(runningSessions.keys()),
             });
         });
+
 
         // ── Daemon control ────────────────────────────────────────────────
 
@@ -1630,254 +1148,4 @@ export async function runDaemon(_args: string[] = []): Promise<number> {
             logError(`server error: ${data.message}`);
         });
     });
-}
-
-export function isPidRunning(pid: number): boolean {
-    if (!Number.isFinite(pid) || pid <= 0) return false;
-    try {
-        process.kill(pid, 0);
-    } catch (err: any) {
-        // ESRCH = process does not exist. EPERM = exists but no permission.
-        if (err?.code === "ESRCH") return false;
-        // On Windows, process.kill(pid, 0) throws EPERM when the process exists
-        // but we lack permission, and throws with code ESRCH (or sometimes just
-        // a generic error) when it doesn't.  If we get here without ESRCH, assume alive.
-    }
-
-    // The PID is alive, but it may have been reused by an unrelated process.
-    // Verify the command line contains a pizzapi / runner signature.
-    try {
-        let cmd: string;
-        if (process.platform === "win32") {
-            // On Windows, use WMIC to inspect the process command line.
-            // wmic is available on Windows 7+ and returns the full command line.
-            cmd = execFileSync(
-                "wmic",
-                ["process", "where", `ProcessId=${pid}`, "get", "CommandLine", "/format:list"],
-                { encoding: "utf-8", timeout: 5000 },
-            ).trim();
-        } else {
-            cmd = execFileSync("ps", ["-p", String(pid), "-o", "command="], { encoding: "utf-8", timeout: 3000 }).trim();
-        }
-        // Match against known runner process patterns:
-        //   - "bun ... runner"          (dev: bun packages/cli/src/index.ts runner)
-        //   - "bun ... daemon.ts"       (dev: direct daemon run)
-        //   - "bun ... _daemon"         (supervisor-spawned child)
-        //   - "pizzapi ... runner"      (production CLI)
-        //   - "node ... runner"         (unlikely but possible)
-        const isRunner =
-            /\brunner\b/i.test(cmd) ||
-            /\bdaemon\b/i.test(cmd) ||
-            /\bpizzapi\b/i.test(cmd) ||
-            /\b_daemon\b/i.test(cmd);
-        if (!isRunner) {
-            // PID exists but belongs to an unrelated process — stale lock.
-            return false;
-        }
-    } catch {
-        // If we can't check the command (e.g. ps/wmic not available), fall back to
-        // assuming the process is the runner (safe default — avoids double-start).
-    }
-
-    return true;
-}
-
-function parseRoots(raw: string): string[] {
-    return raw
-        .split(",")
-        .map((s) => s.trim())
-        .filter(Boolean)
-        .map((s) => s.replace(/\\/g, "/"))
-        .map((s) => (s.length > 1 ? s.replace(/\/+$/, "") : s));
-}
-
-function getWorkspaceRoots(): string[] {
-    // Preferred env vars
-    const rootsRaw = process.env.PIZZAPI_WORKSPACE_ROOTS;
-    const rootSingle = process.env.PIZZAPI_WORKSPACE_ROOT;
-
-    // Back-compat
-    const legacy = process.env.PIZZAPI_RUNNER_ROOTS;
-
-    if (rootsRaw && rootsRaw.trim()) return parseRoots(rootsRaw);
-    if (rootSingle && rootSingle.trim()) return parseRoots(rootSingle);
-    if (legacy && legacy.trim()) return parseRoots(legacy);
-    return [];
-}
-
-function isCwdAllowed(cwd: string | undefined): boolean {
-    if (!cwd) return true;
-    const roots = getWorkspaceRoots();
-    if (roots.length === 0) return true; // unscoped runner
-    // Resolve symlinks + normalize ".." segments to prevent path traversal.
-    // Use realpathSync when the path exists (resolves symlinks), fall back
-    // to resolve() for non-existent paths (still collapses "..").
-    const canonicalize = (p: string) => {
-        try { return realpathSync(p); } catch { return resolve(p); }
-    };
-    const nCwd = canonicalize(cwd).replace(/\\/g, "/").replace(/\/+$/, "") || "/";
-    // Windows paths are case-insensitive
-    const isWin = /^[A-Za-z]:/.test(cwd);
-    return roots.some((root) => {
-        const nRoot = canonicalize(root).replace(/\\/g, "/").replace(/\/+$/, "") || "/";
-        // Special-case filesystem root: everything is under "/"
-        if (nRoot === "/") return true;
-        const rc = isWin ? nCwd.toLowerCase() : nCwd;
-        const rr = isWin ? nRoot.toLowerCase() : nRoot;
-        return rc === rr || rc.startsWith(rr + "/");
-    });
-}
-
-function spawnSession(
-    sessionId: string,
-    apiKey: string,
-    relayUrl: string,
-    requestedCwd: string | undefined,
-    runningSessions: Map<string, RunnerSession>,
-    restartingSessions: Set<string>,
-    onRestartRequested?: () => void,
-    options?: {
-        prompt?: string;
-        model?: { provider: string; id: string };
-        hiddenModels?: string[];
-        agent?: { name: string; systemPrompt?: string; tools?: string; disallowedTools?: string };
-        parentSessionId?: string;
-    },
-): void {
-    logInfo(`spawning headless worker for session ${sessionId}…`);
-
-    if (runningSessions.has(sessionId)) {
-        throw new Error(`Session already running: ${sessionId}`);
-    }
-
-    // Resolve the effective cwd for this session now so we can register it for
-    // usage auth lookups and clean it up on exit without re-deriving it.
-    const effectiveCwd = requestedCwd ?? process.cwd();
-
-    if (!isCwdAllowed(requestedCwd)) {
-        throw new Error(`Requested cwd is outside allowed workspace root(s): ${requestedCwd}`);
-    }
-
-    if (requestedCwd) {
-        if (!existsSync(requestedCwd)) {
-            throw new Error(`cwd does not exist: ${requestedCwd}`);
-        }
-        const st = statSync(requestedCwd);
-        if (!st.isDirectory()) {
-            throw new Error(`cwd is not a directory: ${requestedCwd}`);
-        }
-    }
-
-    const workerArgs = resolveWorkerSpawnArgs();
-
-    const env: Record<string, string> = {
-        ...Object.fromEntries(Object.entries(process.env).filter(([, v]) => typeof v === "string")) as any,
-        // Use the daemon's resolved relay URL so workers always connect to the
-        // same relay the daemon is using (not a potentially-changed config file).
-        PIZZAPI_RELAY_URL: relayUrl,
-        PIZZAPI_API_KEY: apiKey,
-        PIZZAPI_SESSION_ID: sessionId,
-        // Tell the worker where the runner-managed usage cache lives so it can
-        // read quota data without making its own provider API calls.
-        PIZZAPI_RUNNER_USAGE_CACHE_PATH: runnerUsageCacheFilePath(),
-        ...(requestedCwd ? { PIZZAPI_WORKER_CWD: requestedCwd } : {}),
-        // Initial prompt and model for the new session (set by spawn_session tool).
-        ...(options?.prompt ? { PIZZAPI_WORKER_INITIAL_PROMPT: options.prompt } : {}),
-        ...(options?.model ? {
-            PIZZAPI_WORKER_INITIAL_MODEL_PROVIDER: options.model.provider,
-            PIZZAPI_WORKER_INITIAL_MODEL_ID: options.model.id,
-        } : {}),
-        // Hidden model keys (JSON array of "provider/modelId" strings).
-        // The list_models tool filters these from its output.
-        ...(options?.hiddenModels && options.hiddenModels.length > 0
-            ? { PIZZAPI_HIDDEN_MODELS: JSON.stringify(options.hiddenModels) }
-            : {}),
-        // Agent session config — spawn the worker "as" this agent.
-        // Parent session ID for trigger system (child→parent communication).
-        ...(options?.parentSessionId ? { PIZZAPI_WORKER_PARENT_SESSION_ID: options.parentSessionId } : {}),
-        ...(options?.agent?.name ? { PIZZAPI_WORKER_AGENT_NAME: options.agent.name } : {}),
-        ...(options?.agent?.systemPrompt ? { PIZZAPI_WORKER_AGENT_SYSTEM_PROMPT: options.agent.systemPrompt } : {}),
-        ...(options?.agent?.tools ? { PIZZAPI_WORKER_AGENT_TOOLS: options.agent.tools } : {}),
-        ...(options?.agent?.disallowedTools ? { PIZZAPI_WORKER_AGENT_DISALLOWED_TOOLS: options.agent.disallowedTools } : {}),
-    };
-
-    const child = spawn(process.execPath, workerArgs, {
-        env,
-        // Include an IPC channel (fd[3]) so the worker can send a "pre_restart"
-        // message to the daemon before calling process.exit(43).  This lets us
-        // add the sessionId to restartingSessions *before* the process exits and
-        // before the relay's session_ended event (which travels over Socket.IO) can
-        // arrive — closing the race where session_ended beats child.on("exit") and
-        // incorrectly deletes attachments for a still-live restarting session.
-        stdio: ["ignore", "inherit", "inherit", "ipc"],
-    });
-
-    // Pre-restart IPC signal: the worker sends this before calling process.exit(43).
-    // Marking restartingSessions here (synchronously, while the worker is still
-    // alive) guarantees the guard is set before any relay session_ended event arrives.
-    child.on("message", (msg: unknown) => {
-        if (typeof msg === "object" && msg !== null && (msg as Record<string, unknown>).type === "pre_restart") {
-            restartingSessions.add(sessionId);
-            logInfo(`session ${sessionId} signaled pre-restart via IPC`);
-        }
-    });
-
-    // Register this session's cwd so usage fetches can probe its project-local
-    // agentDir override (if any).  Cleaned up on exit below.
-    trackSessionCwd(sessionId, effectiveCwd);
-
-    child.on("exit", (code, signal) => {
-        runningSessions.delete(sessionId);
-        untrackSessionCwd(sessionId, effectiveCwd);
-        logInfo(`session ${sessionId} exited (code=${code}, signal=${signal})`);
-        if (code === 43 && onRestartRequested) {
-            // Restart-in-place: re-spawn immediately without touching attachments.
-            // The session continues under the same ID — files saved to
-            // ~/.pizzapi/session-attachments/{sessionId} must survive the restart.
-            // restartingSessions was already populated via the IPC "pre_restart"
-            // message above; this add is a belt-and-suspenders fallback for the
-            // (unlikely) case where the IPC message was not sent or was lost.
-            restartingSessions.add(sessionId);
-            logInfo(`re-spawning session ${sessionId} (worker restart requested)`);
-            onRestartRequested();
-        } else {
-            // True termination — clean up persisted attachments now.
-            // session_ended will also arrive later but runningSessions will be empty
-            // by then, so this is the reliable cleanup point for spawned sessions.
-            void cleanupSessionAttachments(sessionId).catch(() => {});
-        }
-    });
-
-    runningSessions.set(sessionId, { sessionId, child, startedAt: Date.now(), parentSessionId: options?.parentSessionId });
-    logInfo(`session ${sessionId} worker spawned (pid=${child.pid})`);
-}
-
-/** Is this process running inside a compiled Bun single-file binary? */
-// Detect compiled Bun single-file binary.
-// - Unix: import.meta.url contains "$bunfs"
-// - Windows: import.meta.url contains "~BUN" (drive letter/format varies)
-const isCompiledBinary = import.meta.url.includes("$bunfs") || import.meta.url.includes("~BUN") || import.meta.url.includes("%7EBUN");
-
-/**
- * Returns the spawn arguments for starting a worker subprocess.
- * - Compiled binary: `[process.execPath, ["_worker"]]`
- * - Source / built JS: `[process.execPath, [workerFilePath]]`
- */
-function resolveWorkerSpawnArgs(): string[] {
-    if (isCompiledBinary) {
-        // In a compiled binary, the worker code is embedded. We re-invoke
-        // the same binary with the `_worker` subcommand.
-        return ["_worker"];
-    }
-
-    const ext = import.meta.url.endsWith(".ts") ? "ts" : "js";
-    const url = new URL(`./worker.${ext}`, import.meta.url);
-    const path = fileURLToPath(url);
-    if (!existsSync(path)) {
-        const altExt = ext === "ts" ? "js" : "ts";
-        const alt = fileURLToPath(new URL(`./worker.${altExt}`, import.meta.url));
-        if (existsSync(alt)) return [alt];
-        throw new Error(`Runner worker entrypoint not found: ${path}`);
-    }
-    return [path];
 }
